requirements.txt
sentence-transformers==3.2.2
scikit-learn==1.3.0
google-generativeai==0.8.3
psycopg2==2.9.7
pymongo==4.6.0
numpy==2.2.0
SQLAlchemy==2.0.36